Refactor of project architecture, order on Storage and user service
