+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/evolution/data/di/FirebaseModule.kt
@@ -1,9 +1,9 @@
 package ar.edu.unlam.mobile.scaffolding.evolution.data.di
 
-import ar.edu.unlam.mobile.scaffolding.evolution.data.network.service.StorageService
-import ar.edu.unlam.mobile.scaffolding.evolution.data.network.service.UserDataService
-import ar.edu.unlam.mobile.scaffolding.evolution.data.repository.ImageRepository
-import ar.edu.unlam.mobile.scaffolding.evolution.data.repository.UserDataRepository
+import ar.edu.unlam.mobile.scaffolding.evolution.domain.repository.ImageRepository
+import ar.edu.unlam.mobile.scaffolding.evolution.domain.repository.UserDataRepository
+import ar.edu.unlam.mobile.scaffolding.evolution.domain.services.StorageService
+import ar.edu.unlam.mobile.scaffolding.evolution.domain.services.UserDataService
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ktx.firestore
