+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/evolution/domain/repository/ImageRepository.kt
@@ -1,7 +1,7 @@
-package ar.edu.unlam.mobile.scaffolding.evolution.data.repository
+package ar.edu.unlam.mobile.scaffolding.evolution.domain.repository
 
 import android.net.Uri
-import ar.edu.unlam.mobile.scaffolding.evolution.data.network.model.UserDataResponse
+import ar.edu.unlam.mobile.scaffolding.evolution.domain.model.UserDataResponse
 
 typealias AddImageToStorageResponse = UserDataResponse<Uri>
 typealias AddImageUrlToFirestoreResponse = UserDataResponse<Boolean>
