+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/evolution/ui/screens/uploadImageScreen/image/GetImageFromDatabase.kt
@@ -2,9 +2,9 @@ package ar.edu.unlam.mobile.scaffolding.evolution.ui.screens.uploadImageScreen.i
 
 import androidx.compose.runtime.Composable
 import androidx.hilt.navigation.compose.hiltViewModel
-import ar.edu.unlam.mobile.scaffolding.evolution.data.network.model.UserDataResponse.Failure
-import ar.edu.unlam.mobile.scaffolding.evolution.data.network.model.UserDataResponse.Loading
-import ar.edu.unlam.mobile.scaffolding.evolution.data.network.model.UserDataResponse.Success
+import ar.edu.unlam.mobile.scaffolding.evolution.domain.model.UserDataResponse.Failure
+import ar.edu.unlam.mobile.scaffolding.evolution.domain.model.UserDataResponse.Loading
+import ar.edu.unlam.mobile.scaffolding.evolution.domain.model.UserDataResponse.Success
 import ar.edu.unlam.mobile.scaffolding.evolution.ui.screens.homeLoginProfile.userProfileScreen.viewmodel.UploadImageScreenViewModel
 
 @Composable
